pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "glassnode-python"
version = "0.1.0"
description = "Glassnode API client that mimics the yfinance download experience."
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Leev1s" }]
license = { file = "LICENSE" }
keywords = ["glassnode", "crypto", "api", "pandas"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "pandas>=1.5",
    "requests>=2.31",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
test = ["pytest>=7.4"]

[project.urls]
Homepage = "https://github.com/Leev1s/glassnode-python"
Source = "https://github.com/Leev1s/glassnode-python"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
